+++ jib-core/src/test/java/com/google/cloud/tools/jib/image/json/JsonToImageTranslatorTest.java
@@ -113,26 +113,25 @@ public class JsonToImageTranslatorTest {
 
   @Test
   public void testJsonToImageTranslatorRegex() {
-    assertGoodPattern("NAME=VALUE", "NAME", "VALUE");
-    assertGoodPattern("A1203921=www=ww", "A1203921", "www=ww");
-    assertGoodPattern("abcABC123=", "abcABC123", "");
-    assertGoodPattern("m_a_8943=100", "m_a_8943", "100");
-    assertGoodPattern("A_B_C_D=*****", "A_B_C_D", "*****");
+    assertGoodEnvironmentPattern("NAME=VALUE", "NAME", "VALUE");
+    assertGoodEnvironmentPattern("A1203921=www=ww", "A1203921", "www=ww");
+    assertGoodEnvironmentPattern("&*%(&#$(*@(%&@$*$(=", "&*%(&#$(*@(%&@$*$(", "");
+    assertGoodEnvironmentPattern("m_a_8943=100", "m_a_8943", "100");
+    assertGoodEnvironmentPattern("A_B_C_D=*****", "A_B_C_D", "*****");
 
-    assertBadPattern("123=ABC");
-    assertBadPattern("=================");
-    assertBadPattern("A_B_C");
-    assertBadPattern("$$$$$=fjdkslajfkdsl");
+    assertBadEnvironmentPattern("=================");
+    assertBadEnvironmentPattern("A_B_C");
   }
 
-  private void assertGoodPattern(String input, String expectedName, String expectedValue) {
+  private void assertGoodEnvironmentPattern(
+      String input, String expectedName, String expectedValue) {
     Matcher matcher = JsonToImageTranslator.environmentPattern.matcher(input);
     Assert.assertTrue(matcher.matches());
     Assert.assertEquals(expectedName, matcher.group(1));
     Assert.assertEquals(expectedValue, matcher.group(2));
   }
 
-  private void assertBadPattern(String input) {
+  private void assertBadEnvironmentPattern(String input) {
     Matcher matcher = JsonToImageTranslator.environmentPattern.matcher(input);
     Assert.assertFalse(matcher.matches());
   }