+++ jib-maven-plugin/CHANGELOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- Environment variable configuration is propagated from the base image ([#716](https://github.com/GoogleContainerTools/jib/pull/716))
+- Propagates environment variables from the base image ([#716](https://github.com/GoogleContainerTools/jib/pull/716))
 
 ### Fixed